fix(decidables-elements): fix filter effects region for shadows
The filter effects region must be declared in `userSpaceOnUse` units to
work with straight lines (which have no x or no y dimension). But Chrome
has a bug
(https://bugs.chromium.org/p/chromium/issues/detail?id=294900),
so `userSpaceOnUse` only works when the filter is defined in the same
svg it is used. So the filters must be
inserted into the main svg.