feat(sites): move font config to shared YAML in `decidables-site`
Fonts are now listed in YAML documents. They are read in to both
Sass and JS build scripts. This allows them to be specified once in
`decidables-site` and then used by all sites. It also allows `decidables`
and `dable` to specify additional fonts in a clean way by passing
them in to the `decidables-site` font machinery and the
compile/build scripts.