fix(detectable-elements): fix iso-* contour gradients in `roc-space`
A recent update of SASS lead to the use of floats in `rgb()` colors in CSS. Then
these colors are retrieved in JS and passed to `d3-color` which can't handle
floats in color specs.

In order to properly handle all color values returned from CSS, now using the
`color` package to parse the colors and convert to HEX format safe for `d3-color`.

Fixes #21